soc: Allow switching the core to marocchino
Allow using the --option_pipeline parameter to switch the
cpu core implementation from mor1kx to or1k_maroccino.

This doesn't work for me now as marocchino synth ends up
running out of space on the fpga.  However, the mor1kx
soc still works so having this is a good option to allow
us to easily work on in the future.